refactor:  Repository의 ID 타입을 Long으로 변경 및 반환 타입을 Optional로 개선
- 메서드의 반환 타입을 Optional<Password>로 변경하여 결과가 없을 경우를 처리
